+++ packages/traceloop-sdk/tests/conftest.py
@@ -1,4 +1,6 @@
 """Unit tests configuration module."""
+
+import os
 import pytest
 from traceloop.sdk import Traceloop
 from opentelemetry.sdk.trace.export.in_memory_span_exporter import InMemorySpanExporter
@@ -9,10 +11,25 @@ pytest_plugins = []
 @pytest.fixture(scope="session")
 def exporter():
     exporter = InMemorySpanExporter()
-    Traceloop.init(app_name="test", disable_batch=True, exporter=exporter)
+    Traceloop.init(
+        app_name="test",
+        resource_attributes={"something": "yes"},
+        disable_batch=True,
+        exporter=exporter,
+    )
     return exporter
 
 
 @pytest.fixture(autouse=True)
 def clear_exporter(exporter):
     exporter.clear()
+
+
+@pytest.fixture(autouse=True)
+def environment():
+    os.environ["OPENAI_API_KEY"] = "test_api_key"
+
+
+@pytest.fixture(scope="module")
+def vcr_config():
+    return {"filter_headers": ["authorization"]}